Removes development sample of button at sampe app
