+++ SampleApp/Sources/Sample/Components/ComponentsSection.swift
@@ -8,7 +8,6 @@ class ComponentsSection: SampleSection {
         TextFieldItemViewController.self,
         ValueTextHighlightItemViewController.self,
         ButtonsItemViewController.self,
-        ButtonsItemViewControllerN.self,
         IconographyItemViewController.self,
         NavigationDrawerItemViewController.self,
         DividerViewController.self
